LastFM and Twitter resources now work - no persistence, just stored in memory. Added a service to send email messages to me.

+++ src/main/scala/me/gregd/www/model/Tweet.scala
@@ -8,7 +8,7 @@ import javax.persistence._
 class Tweet (
   
   @Id @Basic(optional = false)
-  @BeanProperty var id:Integer,
+  @BeanProperty var id:Long,
   
   @Column(name="screenname")
   @BeanProperty var screenName:String,
